Archive Pacifica C2 jvePilot setup

+++ selfdrive/test/setup_device_ci.sh
@@ -1,6 +1,11 @@
-#!/usr/bin/bash -e
+#!/usr/bin/bash
 
-export SOURCE_DIR="/data/openpilot_source/"
+set -e
+
+if [ -z "$SOURCE_DIR" ]; then
+  echo "SOURCE_DIR must be set"
+  exit 1
+fi
 
 if [ -z "$GIT_COMMIT" ]; then
   echo "GIT_COMMIT must be set"
@@ -12,29 +17,56 @@ if [ -z "$TEST_DIR" ]; then
   exit 1
 fi
 
-if [ ! -d "$SOURCE_DIR" ]; then
-  git clone https://github.com/commaai/openpilot.git "$SOURCE_DIR"
+umount /data/safe_staging/merged/ || true
+sudo umount /data/safe_staging/merged/ || true
+
+if [ -f "/EON" ]; then
+  rm -rf /data/core
+  rm -rf /data/neoupdate
+  rm -rf /data/safe_staging
 fi
 
-# clear scons cache dirs that haven't been written to in one day
-#cd /tmp && find -name 'scons_cache_*' -type d -maxdepth 1 -mtime +1 -exec rm -rf '{}' \;
+export KEYS_PATH="/usr/comma/setup_keys"
+export CONTINUE_PATH="/data/continue.sh"
+if [ -f "/EON" ]; then
+  export KEYS_PATH="/data/data/com.termux/files/home/setup_keys"
+  export CONTINUE_PATH="/data/data/com.termux/files/continue.sh"
+fi
+tee $CONTINUE_PATH << EOF
+#!/usr/bin/bash
 
-# this can get really big on the CI devices
-rm -rf /data/core
+PARAMS_ROOT="/data/params/d"
+
+while true; do
+  mkdir -p \$PARAMS_ROOT
+  cp $KEYS_PATH \$PARAMS_ROOT/GithubSshKeys
+  echo -n 1 > \$PARAMS_ROOT/SshEnabled
+  sleep 1m
+done
+
+sleep infinity
+EOF
+chmod +x $CONTINUE_PATH
 
 # set up environment
+if [ ! -d "$SOURCE_DIR" ]; then
+  git clone https://github.com/commaai/openpilot.git $SOURCE_DIR
+fi
 cd $SOURCE_DIR
+
+rm -f .git/index.lock
 git reset --hard
 git fetch
 find . -maxdepth 1 -not -path './.git' -not -name '.' -not -name '..' -exec rm -rf '{}' \;
+git fetch --verbose origin $GIT_COMMIT
 git reset --hard $GIT_COMMIT
 git checkout $GIT_COMMIT
 git clean -xdf
 git submodule update --init --recursive
-git submodule foreach --recursive git reset --hard
-git submodule foreach --recursive git clean -xdf
-echo "git checkout took $SECONDS seconds"
+git submodule foreach --recursive "git reset --hard && git clean -xdf"
+
+echo "git checkout done, t=$SECONDS"
 
 rsync -a --delete $SOURCE_DIR $TEST_DIR
 
-echo "$TEST_DIR synced with $GIT_COMMIT, took $SECONDS seconds"
+echo "$TEST_DIR synced with $GIT_COMMIT, t=$SECONDS"
